Rename lib-franklin.js to aem.js #91
