+++ blocks/blog-results/blog-results.js
@@ -1,5 +1,5 @@
 import { getTextLabel, getJsonFromUrl, createElement } from '../../scripts/common.js';
-import { readBlockConfig } from '../../scripts/lib-franklin.js';
+import { readBlockConfig } from '../../scripts/aem.js';
 
 const blockName = 'blog-results';
