+++ blocks/results-list/product-card.js
@@ -1,5 +1,5 @@
 import { getTextLabel, createElement, getLocaleContextedUrl } from '../../scripts/common.js';
-import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
+import { createOptimizedPicture } from '../../scripts/aem.js';
 
 const blockName = 'product-card';
 const partNumberText = getTextLabel('part_number');
